Add websocket connection
Adds a websocket connection to the server so the chat can used from a
browser or from a telnet-style client.

This required the `client` to become an interface, so that each client
implementation could satisfy (ie, tcpClient and wsClient).

Also, to make commands easier to deal with, a funcmap is used now.